refactor: split tournament controller
