+++ backend/src/controllers/tournaments/index.js
@@ -1,0 +1,15 @@
+module.exports = {
+  createTournament: require('./createTournament'),
+  listTournaments: require('./listTournaments'),
+  getTournamentById: require('./getTournamentById'),
+  registerPlayer: require('./registerPlayer'),
+  manualRegisterPlayer: require('./manualRegisterPlayer'),
+  addRebuy: require('./addRebuy'),
+  addAddon: require('./addAddon'),
+  checkInPlayer: require('./checkInPlayer'),
+  eliminatePlayer: require('./eliminatePlayer'),
+  exportTournamentResults: require('./exportTournamentResults'),
+  settleRegistrationPayment: require('./settleRegistrationPayment'),
+  deleteTournament: require('./deleteTournament'),
+  updateRebuyCount: require('./updateRebuyCount')
+};
